+++ packages/anthropic/src/anthropic-messages-settings.ts
@@ -1,5 +1,6 @@
 // https://docs.anthropic.com/claude/docs/models-overview
 export type AnthropicMessagesModelId =
+  | 'claude-3-5-sonnet-20240620'
   | 'claude-3-opus-20240229'
   | 'claude-3-sonnet-20240229'
   | 'claude-3-haiku-20240307'